fix: abortable sleep
